refactor: Account대신 User를 사용
