+++ backend/src/auth/model.py
@@ -5,20 +5,16 @@ from ..model import CreatedAtMixin, UpdatedAtMixin
 
 
 class OAuthAccount(CreatedAtMixin, UpdatedAtMixin, SQLModelBaseOAuthAccount, table=True):
-    __tablename__: str = "oauth_account"
-
     id: int | None = Field(default=None, primary_key=True)
+    user_id: int = Field(foreign_key="user.id")
 
 
-class Account(CreatedAtMixin, UpdatedAtMixin, SQLModelBaseUserDB, table=True):
-    __tablename__: str = "account"
-
+class User(CreatedAtMixin, UpdatedAtMixin, SQLModelBaseUserDB, table=True):
     id: int | None = Field(default=None, primary_key=True)
     email: str = Field(max_length=20)
-    password: str | None = Field()
 
 
-class GuestAccount(CreatedAtMixin, UpdatedAtMixin, SQLModel, table=True):
+class GuestUser(CreatedAtMixin, UpdatedAtMixin, SQLModel, table=True):
     __tablename__: str = "guest_account"
 
     id: int | None = Field(default=None, primary_key=True)
