+++ backend/src/auth/router.py
@@ -2,12 +2,12 @@ from fastapi import APIRouter
 from fastapi_users import FastAPIUsers
 
 from ..config import settings
-from .account import auth_backend
-from .dependency import get_account_manager
-from .model import Account
+from .dependency import get_user_manager
+from .model import User
 from .oauth2 import google_oauth_client
+from .user import auth_backend
 
-fastapi_accounts = FastAPIUsers[Account, int](get_account_manager, [auth_backend])
+fastapi_accounts = FastAPIUsers[User, int](get_user_manager, [auth_backend])
 
 router = APIRouter(tags=["auth"])
 # router.include_router(fastapi_accounts.get_auth_router(auth_backend), prefix="/auth")
