+++ backend/src/auth/user.py
@@ -3,14 +3,14 @@ from fastapi_users import BaseUserManager, IntegerIDMixin
 from fastapi_users.authentication import AuthenticationBackend, BearerTransport, JWTStrategy
 
 from ..config import settings
-from .model import Account
+from .model import User
 
 
-class AccountManager(IntegerIDMixin, BaseUserManager[Account, int]):
+class UserManager(IntegerIDMixin, BaseUserManager[User, int]):
     reset_password_token_secret = settings.JWT_SECRET
     verification_token_secret = settings.JWT_SECRET
 
-    async def on_after_register(self, account: Account, request: Request | None = None) -> None:
+    async def on_after_register(self, account: User, request: Request | None = None) -> None:
         ...
